# Conflicts:
#	Parameters.json

+++ teamup-input/src/main/java/ru/team/up/input/controller/privateController/EventController.java
@@ -17,10 +17,12 @@ import ru.team.up.core.mappers.EventMapper;
 import ru.team.up.core.monitoring.service.MonitorProducerService;
 import ru.team.up.core.service.EventService;
 import ru.team.up.dto.*;
+import ru.team.up.sup.service.ParameterService;
 
 import javax.persistence.PersistenceException;
 import javax.validation.constraints.NotNull;
 import java.util.HashMap;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -53,13 +55,24 @@ public class EventController {
     @GetMapping
     public ResponseEntity<List<EventDto>> getAllEvents() {
         log.debug("Старт метода ResponseEntity<List<EventDto>> getAllEvents()");
+        if (!ParameterService.getAllEventsPrivateEnabled.getValue()) {
+            log.debug("Метод getAllEvents выключен параметром getAllEventsPrivateEnabled = false");
+            throw new RuntimeException("Method getAllEvents is disabled by parameter getAllEventsPrivateEnabled");
+        }
         List<Event> events = eventService.getAllEvents();
         ResponseEntity<List<EventDto>> responseEntity = ResponseEntity.ok(
                 EventMapper.INSTANCE.mapDtoEventToEvent(events));
         log.debug("Сформирован ответ {}", responseEntity);
-        Map<String, Object> monitoringParameters = new HashMap<>();
-        monitoringParameters.put("Количество всех мероприятий ",
-                events.size());
+
+        Map<String, ParametersDto> monitoringParameters = new HashMap<>();
+
+        ParametersDto eventsSize = ParametersDto.builder()
+                .description("Количество всех мероприятий")
+                .value(events.size())
+                .build();
+
+        monitoringParameters.put("Количество всех мероприятий ", eventsSize);
+
         monitoringProducerService.send(
                 monitoringProducerService.constructReportDto(
                         SecurityContextHolder.getContext().getAuthentication().getPrincipal(), ControlDto.MANUAL,
@@ -81,14 +94,30 @@ public class EventController {
     @GetMapping("/{id}")
     public ResponseEntity<EventDto> getOneEvent(@PathVariable Long id) {
         log.debug("Старт метода ResponseEntity<EventDto> getOneEvent(@PathVariable Long id) с параметром {}", id);
+        if (!ParameterService.getOneEventEnabled.getValue()) {
+            log.debug("Метод getOneEvent выключен параметром getOneEventEnabled = false");
+            throw new RuntimeException("Method getOneEvent is disabled by parameter getOneEventEnabled");
+        }
         Event event = eventService.getOneEvent(id);
         ResponseEntity<EventDto> responseEntity = ResponseEntity.ok(
                 EventMapper.INSTANCE.mapEventToDto(event)
         );
         log.debug("Сформирован ответ {}", responseEntity);
-        Map<String, Object> monitoringParameters = new HashMap<>();
-        monitoringParameters.put("ID мероприятия ", event.getId());
-        monitoringParameters.put("Название мероприятия ", event.getEventName());
+
+        Map<String, ParametersDto> monitoringParameters = new LinkedHashMap<>();
+
+        ParametersDto oneEventId = ParametersDto.builder()
+                .description("Id мероприятия ")
+                .value(event.getId())
+                .build();
+
+        ParametersDto oneEventName = ParametersDto.builder()
+                .description("Название мероприятия ")
+                .value(event.getEventName())
+                .build();
+
+        monitoringParameters.put("Id мероприятия ", oneEventId);
+        monitoringParameters.put("Название мероприятия ", oneEventName);
         monitoringProducerService.send(
                 monitoringProducerService.constructReportDto(
                         SecurityContextHolder.getContext().getAuthentication().getPrincipal(), ControlDto.MANUAL,
@@ -110,14 +139,30 @@ public class EventController {
     @GetMapping("viewEvent/{id}")
     public ResponseEntity<EventDto> updateNumberOfParticipants(@PathVariable Long id) {
         log.debug("Старт метода ResponseEntity<Event> updateNumberOfViews(@PathVariable Long id) с параметром {}", id);
+        if (!ParameterService.updateNumberOfParticipantsEnabled.getValue()) {
+            log.debug("Метод updateNumberOfParticipants выключен параметром updateNumberOfParticipantsEnabled = false");
+            throw new RuntimeException("Method updateNumberOfParticipants is disabled by parameter updateNumberOfParticipantsEnabled");
+        }
         Event event = eventService.getOneEvent(id);
         eventService.updateNumberOfViews(id);
         ResponseEntity<EventDto> responseEntity = new ResponseEntity<>(
                 EventMapper.INSTANCE.mapEventToDto(eventService.getOneEvent(id)), HttpStatus.ACCEPTED);
         log.debug("Сформирован ответ {}", responseEntity);
-        Map<String, Object> monitoringParameters = new HashMap<>();
-        monitoringParameters.put("ID мероприятия ", event.getId());
-        monitoringParameters.put("Название мероприятия ", event.getEventName());
+
+        Map<String, ParametersDto> monitoringParameters = new LinkedHashMap<>();
+
+        ParametersDto eventId = ParametersDto.builder()
+                .description("Id мероприятия ")
+                .value(event.getId())
+                .build();
+
+        ParametersDto eventName = ParametersDto.builder()
+                .description("Название мероприятия ")
+                .value(event.getEventName())
+                .build();
+
+        monitoringParameters.put("Id мероприятия ", eventId);
+        monitoringParameters.put("Название мероприятия ", eventName);
         monitoringProducerService.send(
                 monitoringProducerService.constructReportDto(
                         SecurityContextHolder.getContext().getAuthentication().getPrincipal(), ControlDto.MANUAL,
@@ -146,9 +191,34 @@ public class EventController {
     @PostMapping
     public ResponseEntity<Event> createEvent(@RequestBody @NotNull Event eventCreate) {
         log.debug("Старт метода ResponseEntity<Event> createEvent(@RequestBody @NotNull Event event) с параметром {}", eventCreate);
-
+        if (!ParameterService.createEventEnabled.getValue()) {
+            log.debug("Метод createEvent выключен параметром createEventEnabled = false");
+            throw new RuntimeException("Method createEvent is disabled by parameter createEventEnabled");
+        }
         try {
             ResponseEntity<Event> responseEntity = new ResponseEntity<>(eventService.saveEvent(eventCreate), HttpStatus.CREATED);
+
+            Map<String, ParametersDto> monitoringParameters = new LinkedHashMap<>();
+
+            ParametersDto eventId = ParametersDto.builder()
+                    .description("Id мероприятия ")
+                    .value(eventCreate.getId())
+                    .build();
+
+            ParametersDto eventName = ParametersDto.builder()
+                    .description("Название мероприятия ")
+                    .value(eventCreate.getEventName())
+                    .build();
+
+            monitoringParameters.put("Id мероприятия ", eventId);
+            monitoringParameters.put("Название мероприятия ", eventName);
+
+
+            monitoringProducerService.send(
+                    monitoringProducerService.constructReportDto(
+                            SecurityContextHolder.getContext().getAuthentication().getPrincipal(), ControlDto.MANUAL,
+                            this.getClass(), monitoringParameters));
+
             log.debug("Сформирован ответ {}", responseEntity);
             return responseEntity;
         } catch (PersistenceException e) {
@@ -162,7 +232,7 @@ public class EventController {
     }
 
     /**
-     * @param id id обновляемого ивента
+     * @param id    id обновляемого ивента
      * @param event Обновляемый объект класса Event
      * @return Результат работы метода eventService.saveEvent(event)) в виде объекта Event
      * в теле ResponseEntity
@@ -183,7 +253,10 @@ public class EventController {
     @PutMapping("/{id}")
     public ResponseEntity<Event> updateEvent(@PathVariable Long id, @RequestBody @NotNull Event event) {
         log.debug("Старт метода ResponseEntity<Event> updateEvent(@RequestBody @NotNull Event event) с параметром {}", event);
-
+        if (!ParameterService.updateEventEnabled.getValue()) {
+            log.debug("Метод updateEvent выключен параметром updateEventEnabled = false");
+            throw new RuntimeException("Method updateEvent is disabled by parameter updateEventEnabled");
+        }
         if (!id.equals(event.getId())) {
             log.warn("Введен некорректный id");
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -191,7 +264,29 @@ public class EventController {
 
         try {
             ResponseEntity<Event> responseEntity = ResponseEntity.ok(eventService.updateEvent(event));
+
+            Map<String, ParametersDto> monitoringParameters = new LinkedHashMap<>();
+
+            ParametersDto eventId = ParametersDto.builder()
+                    .description("Id мероприятия ")
+                    .value(event.getId())
+                    .build();
+
+            ParametersDto eventName = ParametersDto.builder()
+                    .description("Название мероприятия ")
+                    .value(event.getEventName())
+                    .build();
+
+            monitoringParameters.put("Id мероприятия ", eventId);
+            monitoringParameters.put("Название мероприятия ", eventName);
+
+            monitoringProducerService.send(
+                    monitoringProducerService.constructReportDto(
+                            SecurityContextHolder.getContext().getAuthentication().getPrincipal(), ControlDto.MANUAL,
+                            this.getClass(), monitoringParameters));
+
             log.debug("Сформирован ответ {}", responseEntity);
+
             return responseEntity;
         } catch (PersistenceException e) {
             ResponseEntity<Event> responseEntity = new ResponseEntity<>(null, HttpStatus.BAD_REQUEST);
@@ -210,11 +305,38 @@ public class EventController {
     @DeleteMapping("/{id}")
     public ResponseEntity<Event> deleteAdmin(@PathVariable Long id) {
         log.debug("Старт метода ResponseEntity<Event> deleteAdmin(@PathVariable Long id) с параметром {}", id);
+        if (!ParameterService.deleteAdminEnabled.getValue()) {
+            log.debug("Метод deleteAdmin выключен параметром deleteAdminEnabled = false");
+            throw new RuntimeException("Method deleteAdmin is disabled by parameter deleteAdminEnabled");
+        }
+
+        Event event = eventService.getOneEvent(id);
 
         eventService.deleteEvent(id);
+        log.debug("Событие с id = {} было успешно удалено", id);
 
         ResponseEntity<Event> responseEntity = new ResponseEntity<>(HttpStatus.ACCEPTED);
         log.debug("Сформирован ответ {}", responseEntity);
+
+        Map<String, ParametersDto> monitoringParameters = new LinkedHashMap<>();
+
+        ParametersDto eventId = ParametersDto.builder()
+                .description("Id мероприятия ")
+                .value(event.getId())
+                .build();
+
+        ParametersDto eventName = ParametersDto.builder()
+                .description("Название мероприятия ")
+                .value(event.getEventName())
+                .build();
+
+        monitoringParameters.put("Id мероприятия ", eventId);
+        monitoringParameters.put("Название мероприятия ", eventName);
+
+        monitoringProducerService.send(
+                monitoringProducerService.constructReportDto(
+                        SecurityContextHolder.getContext().getAuthentication().getPrincipal(), ControlDto.MANUAL,
+                        this.getClass(), monitoringParameters));
 
         return responseEntity;
     }
