+++ teamup-input/src/main/java/ru/team/up/input/controller/privateController/ModeratorController.java
@@ -7,46 +7,74 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import ru.team.up.core.entity.Account;
 import ru.team.up.core.entity.Event;
 import ru.team.up.core.entity.Moderator;
+import ru.team.up.core.monitoring.service.MonitorProducerService;
 import ru.team.up.core.service.AssignedEventsService;
 import ru.team.up.core.service.ModeratorService;
+import ru.team.up.dto.ControlDto;
+import ru.team.up.dto.ParametersDto;
+import ru.team.up.sup.service.ParameterService;
 
 import javax.validation.constraints.NotNull;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @author Alexey Tkachenko
- *
  * @link localhost:8080/swagger-ui.html
- * @link  http://localhost:8080/swagger-ui/index.html?configUrl=/v3/api-docs/swagger-config
+ * @link http://localhost:8080/swagger-ui/index.html?configUrl=/v3/api-docs/swagger-config
  * Документация API
  */
 
 @Slf4j
-@Tag(name = "Moderator Private Controller",description = "Moderator API")
+@Tag(name = "Moderator Private Controller", description = "Moderator API")
 @RestController
 @AllArgsConstructor(onConstructor = @__(@Autowired))
 @RequestMapping("/private/account/moderator")
 public class ModeratorController {
+
     private ModeratorService moderatorService;
     private AssignedEventsService assignedEventsService;
-
+    private MonitorProducerService monitorProducerService;
 
     /**
      * @return Результат работы метода moderatorService.getAllModerators() в виде коллекции модераторов
      * в теле ResponseEntity
      */
-    @Operation(summary ="Получение списка всех модераторов")
+    @Operation(summary = "Получение списка всех модераторов")
     @GetMapping
     public ResponseEntity<List<Account>> getAllModerators() {
         log.debug("Старт метода ResponseEntity<List<Moderator>> getAllModerators()");
+        if (!ParameterService.getAllModeratorsEnabled.getValue()) {
+            log.debug("Метод getAllModerators выключен параметром getAllModeratorsEnabled = false");
+            throw new RuntimeException("Method getAllModerators is disabled by parameter getAllModeratorsEnabled");
+        }
 
-        ResponseEntity<List<Account>> responseEntity = ResponseEntity.ok(moderatorService.getAllModerators());
+        List<Account> allModerators = moderatorService.getAllModerators();
+
+        ResponseEntity<List<Account>> responseEntity = ResponseEntity.ok(allModerators);
         log.debug("Получили ответ {}", responseEntity);
 
+        Map<String, ParametersDto> monitoringParameters = new HashMap<>();
+
+        ParametersDto allModeratorsParam = ParametersDto.builder()
+                .description("Количество модераторов ")
+                .value(allModerators.size())
+                .build();
+
+        monitoringParameters.put("Количество модераторов", allModeratorsParam);
+
+        monitorProducerService.send(
+                monitorProducerService.constructReportDto(SecurityContextHolder.getContext().getAuthentication().getPrincipal(),
+                        ControlDto.MANUAL,
+                        this.getClass(),
+                        monitoringParameters)
+        );
         return responseEntity;
     }
 
@@ -55,14 +83,46 @@ public class ModeratorController {
      * @return Результат работы метода moderatorService.getOneModerator(id) в виде объекта Moderator
      * в теле ResponseEntity
      */
-    @Operation(summary ="Получение модератора по id")
+    @Operation(summary = "Получение модератора по id")
     @GetMapping("/{id}")
     public ResponseEntity<Account> getOneModerator(@PathVariable Long id) {
         log.debug("Старт метода ResponseEntity<Moderator> getOneModerator(@PathVariable Long id) с параметром {}", id);
+        if (!ParameterService.getOneModeratorEnabled.getValue()) {
+            log.debug("Метод getOneModerator выключен параметром getOneModeratorEnabled = false");
+            throw new RuntimeException("Method getOneModerator is disabled by parameter getOneModeratorEnabled");
+        }
 
-        ResponseEntity<Account> responseEntity = ResponseEntity.ok(moderatorService.getOneModerator(id));
+        Account moderator = moderatorService.getOneModerator(id);
+        ResponseEntity<Account> responseEntity = ResponseEntity.ok(moderator);
         log.debug("Получили ответ {}", responseEntity);
 
+        Map<String, ParametersDto> monitoringParameters = new HashMap<>();
+
+        ParametersDto moderatorId = ParametersDto.builder()
+                .description("ID")
+                .value(moderator.getId())
+                .build();
+
+        ParametersDto moderatorEmail = ParametersDto.builder()
+                .description("Email")
+                .value(moderator.getEmail())
+                .build();
+
+        ParametersDto moderatorName = ParametersDto.builder()
+                .description("Имя")
+                .value(moderator.getUsername())
+                .build();
+
+        monitoringParameters.put("ID", moderatorId);
+        monitoringParameters.put("Email", moderatorEmail);
+        monitoringParameters.put("Имя", moderatorName);
+
+        monitorProducerService.send(
+                monitorProducerService.constructReportDto(SecurityContextHolder.getContext().getAuthentication().getPrincipal(),
+                        ControlDto.MANUAL,
+                        this.getClass(),
+                        monitoringParameters)
+        );
         return responseEntity;
     }
 
@@ -70,33 +130,98 @@ public class ModeratorController {
      * @return Результат работы метода moderatorService.saveModerator(moderator) в виде объекта Moderator
      * в теле ResponseEntity
      */
-    @Operation(summary ="Создание нового модератора")
+    @Operation(summary = "Создание нового модератора")
     @PostMapping
     public ResponseEntity<Account> createModerator(@RequestBody @NotNull Account moderatorCreate) {
         log.debug("Старт метода ResponseEntity<Moderator> createModerator(@RequestBody @NotNull Moderator moderator) с параметром {}", moderatorCreate);
+        if (!ParameterService.createModeratorEnabled.getValue()) {
+            log.debug("Метод createModerator выключен параметром createModeratorEnabled = false");
+            throw new RuntimeException("Method createModerator is disabled by parameter createModeratorEnabled");
+        }
         ResponseEntity<Account> responseEntity =
                 new ResponseEntity<>(moderatorService.saveModerator(moderatorCreate), HttpStatus.CREATED);
 
         log.debug("Получили ответ {}", responseEntity);
+
+        Map<String, ParametersDto> monitoringParameters = new HashMap<>();
+
+        ParametersDto moderatorId = ParametersDto.builder()
+                .description("ID")
+                .value(moderatorCreate.getId())
+                .build();
+
+        ParametersDto moderatorEmail = ParametersDto.builder()
+                .description("Email")
+                .value(moderatorCreate.getEmail())
+                .build();
+
+        ParametersDto moderatorName = ParametersDto.builder()
+                .description("Имя")
+                .value(moderatorCreate.getUsername())
+                .build();
+
+        monitoringParameters.put("ID", moderatorId);
+        monitoringParameters.put("Email", moderatorEmail);
+        monitoringParameters.put("Имя", moderatorName);
+
+        monitorProducerService.send(
+                monitorProducerService.constructReportDto(SecurityContextHolder.getContext().getAuthentication().getPrincipal(),
+                        ControlDto.MANUAL,
+                        this.getClass(),
+                        monitoringParameters)
+        );
         return responseEntity;
     }
 
     /**
-     * @param moderator Обновляемый объект класса Moderator
+     * @param moderator   Обновляемый объект класса Moderator
      * @param moderatorId id модератора
      * @return Результат работы метода moderatorService.saveModerator(moderator) в виде объекта Moderator
      * в теле ResponseEntity
      */
-    @Operation(summary ="Обновление данных модератора")
+    @Operation(summary = "Обновление данных модератора")
     @PutMapping("/{id}")
-    public ResponseEntity<Moderator> updateModerator(@RequestBody @NotNull Moderator moderator,@PathVariable("id") Long moderatorId) {
+    public ResponseEntity<Moderator> updateModerator(@RequestBody @NotNull Moderator moderator, @PathVariable("id") Long moderatorId) {
         log.debug("Старт метода ResponseEntity<Moderator> updateModerator(@RequestBody @NotNull Moderator moderator) с параметром {}", moderator);
+        if (!ParameterService.updateModeratorEnabled.getValue()) {
+            log.debug("Метод updateModerator выключен параметром updateModeratorEnabled = false");
+            throw new RuntimeException("Method updateModerator is disabled by parameter updateModeratorEnabled");
+        }
         ResponseEntity<Moderator> responseEntity;
         if (moderatorService.moderatorIsExistsById(moderatorId) && moderator.getId().equals(moderatorId)) {
             responseEntity = ResponseEntity.ok(moderatorService.updateModerator(moderator));
             log.debug("Модератор обновлён {}", responseEntity);
+
+            Map<String, ParametersDto> monitoringParameters = new HashMap<>();
+
+            ParametersDto moderatorIdUpdate = ParametersDto.builder()
+                    .description("ID")
+                    .value(moderator.getId())
+                    .build();
+
+            ParametersDto moderatorEmailUpdate = ParametersDto.builder()
+                    .description("Email")
+                    .value(moderator.getEmail())
+                    .build();
+
+            ParametersDto moderatorNameUpdate = ParametersDto.builder()
+                    .description("Имя")
+                    .value(moderator.getUsername())
+                    .build();
+
+            monitoringParameters.put("ID", moderatorIdUpdate);
+            monitoringParameters.put("Email", moderatorEmailUpdate);
+            monitoringParameters.put("Имя", moderatorNameUpdate);
+
+            monitorProducerService.send(
+                    monitorProducerService.constructReportDto(SecurityContextHolder.getContext().getAuthentication().getPrincipal(),
+                            ControlDto.MANUAL,
+                            this.getClass(),
+                            monitoringParameters)
+            );
         } else {
             responseEntity = new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            log.debug("Неверно указан id");
         }
         return responseEntity;
     }
@@ -105,16 +230,34 @@ public class ModeratorController {
      * @param id Удаляемого объекта класса Moderator
      * @return Объект ResponseEntity со статусом OK
      */
-    @Operation(summary ="Удаление модератора по id")
+    @Operation(summary = "Удаление модератора по id")
     @DeleteMapping("/{id}")
     public ResponseEntity<Moderator> deleteModerator(@PathVariable Long id) {
         log.debug("Старт метода ResponseEntity<Moderator> deleteModerator(@PathVariable Long id) с параметром {}", id);
+        if (!ParameterService.deleteModeratorEnabled.getValue()) {
+            log.debug("Метод deleteModerator выключен параметром deleteModeratorEnabled = false");
+            throw new RuntimeException("Method deleteModerator is disabled by parameter deleteModeratorEnabled");
+        }
 
         moderatorService.deleteModerator(id);
 
         ResponseEntity<Moderator> responseEntity = new ResponseEntity<>(HttpStatus.ACCEPTED);
         log.debug("Получили ответ {}", responseEntity);
 
+        Map<String, ParametersDto> monitoringParameters = new HashMap<>();
+
+        ParametersDto deleteModeratorId = ParametersDto.builder()
+                .description("ID")
+                .value(id)
+                .build();
+        monitoringParameters.put("ID", deleteModeratorId);
+
+        monitorProducerService.send(
+                monitorProducerService.constructReportDto(SecurityContextHolder.getContext().getAuthentication().getPrincipal(),
+                        ControlDto.MANUAL,
+                        this.getClass(),
+                        monitoringParameters)
+        );
         return responseEntity;
     }
 
@@ -123,17 +266,35 @@ public class ModeratorController {
      * @return Результат работы метода assignedEventsService.getAllEventsByModeratorId(id) в виде
      * объекта List<AssignedEvents> в теле ResponseEntity
      */
-    @Operation(summary ="Получение мероприятий на проверке модератора по его id")
+    @Operation(summary = "Получение мероприятий на проверке модератора по его id")
     @GetMapping("/{id}/events")
     public ResponseEntity<List<Event>> getAssignedEventsOfModerator(@PathVariable Long id) {
-
         log.debug("Старт метода ResponseEntity<List<Event>> getAssignedEventsOfModerator(@PathVariable Long id)" +
                 " с параметром {}", id);
 
-        ResponseEntity<List<Event>> responseEntity = ResponseEntity.ok(assignedEventsService.getAllEventsByModeratorId(id));
+        if (!ParameterService.getAssignedEventsOfModeratorEnabled.getValue()) {
+            log.debug("Метод getAssignedEventsOfModerator выключен параметром getAssignedEventsOfModeratorEnabled = false");
+            throw new RuntimeException("Method getAssignedEventsOfModerator is disabled by parameter getAssignedEventsOfModeratorEnabled");
+        }
 
+        List<Event> allEvents = assignedEventsService.getAllEventsByModeratorId(id);
+        ResponseEntity<List<Event>> responseEntity = ResponseEntity.ok(allEvents);
         log.debug("Получили ответ {}", responseEntity);
 
+        Map<String, ParametersDto> monitoringParameters = new HashMap<>();
+
+        ParametersDto allEventsOfModerator = ParametersDto.builder()
+                .description("Количество мероприятий ")
+                .value(allEvents.size())
+                .build();
+        monitoringParameters.put("Количество мероприятий", allEventsOfModerator);
+
+        monitorProducerService.send(
+                monitorProducerService.constructReportDto(SecurityContextHolder.getContext().getAuthentication().getPrincipal(),
+                        ControlDto.MANUAL,
+                        this.getClass(),
+                        monitoringParameters)
+        );
         return responseEntity;
     }
 }
